Removed Build model and renamed BuildDetailsInfo to Build
